Move language picker to shared (#11971)
* Begin moving code for language selection out of diagnostic and into shared.

* Move language picker to shared components

* Import each language

* Remove unused lines

---------
