+++ services/QuillLMS/client/app/bundles/Diagnostic/components/eslDiagnostic/footer.tsx
@@ -1,8 +1,7 @@
 import * as React from 'react';
 
 import useWindowSize from '../../../Shared/hooks/useWindowSize';
-import { DropdownInput } from '../../../Shared/index';
-import { languageData, languages } from '../../modules/translation/languagePageInfo';
+import LanguagePicker from '../../../Shared/components/translations/languagePicker';
 
 interface FooterProps {
   diagnosticID: string,
@@ -11,50 +10,25 @@ interface FooterProps {
   handleClickOpenMobileLanguageMenu(): any
 }
 
-const languageIconSrc = `${process.env.CDN_URL}/images/icons/language.svg`
-
 const MAX_VIEW_WIDTH_FOR_MOBILE = 895
-
-const options = (): Array<{value: string, label: string}> => {
-  return languages.map(language => ({
-    value: language,
-    label: (`<p><img alt={${languageData[language].label} flag} src=${languageData[language].flag} /><span>${languageData[language].label}</span></p>`)
-  }))
-}
 
 const Footer = ({ language, handleClickOpenMobileLanguageMenu, updateLanguage, }: FooterProps) => {
   const size = useWindowSize();
-  const onMobile = () => size.width <= MAX_VIEW_WIDTH_FOR_MOBILE
-
-  const onChange = (option: { value: string}) => {
-    const language = option.value;
-    updateLanguage(language)
-  }
-
-  const value = options().find(opt => language === opt.value)
-  if (onMobile()) {
-    return (
-      <div className="ell-footer mobile">
-        <div className="student-container">
-          <button className="passthrough-button focus-on-light" onClick={handleClickOpenMobileLanguageMenu} type="button">
-            <img alt="" src={languageIconSrc} />
-            <span>Change directions language</span>
-          </button>
-        </div>
-      </div>
-    )
+  const footerClassName = () => {
+    if (size.width <= MAX_VIEW_WIDTH_FOR_MOBILE) {
+      return "ell-footer mobile"
+    } else {
+      return "ell-footer"
+    }
   }
 
   return (
-    <div className="ell-footer">
+    <div className={footerClassName()}>
       <div className="student-container">
-        <DropdownInput
-          className="ell-language-selector"
-          handleChange={onChange}
-          label="Directions language"
-          options={options()}
-          usesCustomOption
-          value={value}
+        <LanguagePicker
+          handleClickOpenMobileLanguageMenu={handleClickOpenMobileLanguageMenu}
+          language={language}
+          updateLanguage={updateLanguage}
         />
       </div>
     </div>
